rename package name because directory name changed
