+++ cmd/feature/config_test.go
@@ -1,1 +1,1 @@
-package cmd
+package feature
